refactoring AddPoiService and PoiService into PoiApi. Add PoiRequest model
