+++ BerkutPoiService/Models/PoiRequest.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BerkutPoiService.Models
+{
+    public class PoiRequest
+    {
+        [Required]
+        [Range(-90, 90, ErrorMessage = "Latitude must be between -90 and 90.")]
+        public double? Lat { get; set; }
+
+        [Required]
+        [Range(-180, 180, ErrorMessage = "Longitude must be between -180 and 180.")]
+        public double? Long { get; set; }
+
+        [Required]
+        public string Name { get; set; }
+
+        public string Content { get; set; }
+    }
+}
